Create separate component for the entiities table SN-367
