+++ src/Saritasa.NetForge.Blazor/Controls/Entities.razor.cs
@@ -1,0 +1,11 @@
+﻿using Saritasa.NetForge.Blazor.Pages;
+using Saritasa.NetForge.Mvvm.ViewModels;
+
+namespace Saritasa.NetForge.Blazor.Controls;
+
+/// <summary>
+/// Entities table component.
+/// </summary>
+public partial class Entities : MvvmComponentBase<EntitiesViewModel>
+{
+}
