+++ src/Saritasa.NetForge.Blazor/Pages/Index.razor
@@ -1,0 +1,4 @@
+﻿@using Saritasa.NetForge.Blazor.Controls
+
+<PageTitle>NetForge Admin - Index</PageTitle>
+<Entities/>
