+++ src/Saritasa.NetForge.Blazor/Pages/_Layout.cshtml
@@ -12,6 +12,7 @@
     <link href="_content/NetForgeBlazor/NetForgeBlazor.bundle.scp.css" rel="stylesheet"/>
     <link href="_content/NetForgeBlazor/styles.css" rel="stylesheet"/>
     <component type="typeof(HeadOutlet)" render-mode="ServerPrerendered" />
+    <title>NetForge Admin</title>
 </head>
 <body>
     @RenderBody()
