--- src/Saritasa.NetForge.Blazor/wwwroot/styles.css
@@ -1,14 +1,0 @@
-﻿:root {
-    --grey-hint: #757575;
-}
-
-.text--color-hint {
-    color: var(--grey-hint);
-}
-
-.entities-table__description-text {
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    max-width: 100ch;
-}
