--- src/Saritasa.NetForge.Mvvm/ViewModels/Metadata/MainModel.cs
@@ -1,14 +1,0 @@
-﻿using Saritasa.NetForge.UseCases.Metadata.DTOs;
-
-namespace Saritasa.NetForge.Mvvm.ViewModels.Metadata;
-
-/// <summary>
-/// Model for the index page.
-/// </summary>
-public class MainModel
-{
-    /// <summary>
-    /// Collection of entities metadata.
-    /// </summary>
-    public IEnumerable<EntityMetadataDto> EntitiesMetadata { get; set; } = new List<EntityMetadataDto>();
-}
